Move InputValues to shared types
