+++ src/steps/updateLocalFiles.ts
@@ -1,7 +1,7 @@
 import replaceInFile from "replace-in-file";
 
-import { InputValues } from "../shared/options/readOptions.js";
 import { readFileSafeAsJson } from "../shared/readFileSafeAsJson.js";
+import { InputValues } from "../shared/types.js";
 
 interface ExistingPackageData {
 	description?: string;
